# docs/openapi.yml
openapi: 3.0.3
info:
  title: RIMAC Medical Appointment API
  description: |
    Sistema serverless completo de gestión de citas médicas para RIMAC.
    
    ## Características
    - ✅ Procesamiento asíncrono por país (PE/CL)
    - ✅ Actualizaciones en tiempo real con EventBridge
    - ✅ Panel de administración con JWT
    - ✅ Arquitectura event-driven escalable
    - ✅ Base de datos dual: DynamoDB + RDS MySQL
    - ✅ Dashboard con estadísticas en tiempo real
    
    ## Flujo del Sistema
    1. **Cliente** crea cita → DynamoDB (pending)
    2. **SNS/SQS** → Procesamiento por país
    3. **Lambda específico** → RDS save (pending)
    4. **EventBridge** → Confirmación automática
    5. **Status update** → RDS (completed)
    6. **Admin Dashboard** → Estadísticas en tiempo real
    
  version: 2.0.0
  contact:
    name: RIMAC API Support
    email: api-support@rimac.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://bnh6zvyh59.execute-api.us-east-2.amazonaws.com/dev
    description: Desarrollo
  - url: https://api-prod.rimac.com
    description: Producción
...
paths:
...
  /admin/auth:
    post:
      summary: Autenticación del administrador
      description: |
        Autentica al administrador y retorna un JWT token válido por 24 horas.
        
        **Credenciales de demo:**
        - Usuario: `admin`
        - Contraseña: `rimac2024`
      operationId: adminAuth
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
            example:
              username: "admin"
              password: "rimac2024"
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/stats:
    get:
      summary: Estadísticas del sistema
      description: |
        Obtiene estadísticas completas del sistema incluyendo:
        - Total de citas por status
        - Distribución por país (PE/CL)  
        - Métricas de rendimiento
        - Estado del sistema
      operationId: getAdminStats
      tags:
        - Admin
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: Estadísticas obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatsResponse'
        '401':
          description: Token JWT inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/appointments:
    get:
      summary: Gestión de todas las citas
      description: |
        Obtiene todas las citas del sistema desde RDS (PE + CL) con datos detallados:
        - Información del médico
        - Especialidad médica
        - Centro médico
        - Status actualizado
        
        **Limitado a 50 registros por país para performance.**
      operationId: getAllAppointments
      tags:
        - Admin
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: Citas obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAppointmentsResponse'
        '401':
          description: Token JWT inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/medics:
    get:
      summary: Gestión de médicos
      description: |
        Obtiene todos los médicos del sistema organizados por país con:
        - Información personal del médico
        - Especialidad médica
        - Centro médico asignado
        - Dirección del centro
      operationId: getAllMedics
      tags:
        - Admin
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: Médicos obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMedicsResponse'
        '401':
          description: Token JWT inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/centers:
    get:
      summary: Gestión de centros médicos
      description: |
        Obtiene todos los centros médicos disponibles organizados por país.
      operationId: getAllCenters
      tags:
        - Admin
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: Centros obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCentersResponse'
        '401':
          description: Token JWT inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/specialties:
    get:
      summary: Gestión de especialidades médicas
      description: |
        Obtiene todas las especialidades médicas disponibles organizadas por país.
      operationId: getAllSpecialties
      tags:
        - Admin
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: Especialidades obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSpecialtiesResponse'
        '401':
          description: Token JWT inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# ============= COMPONENTS =============
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtenido desde `/admin/auth`.
        
        **Formato:** `Bearer <token>`
  schemas:
...
    AdminLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Nombre de usuario del administrador
          example: "admin"
        password:
          type: string
          description: Contraseña del administrador
          example: "rimac2024"
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          description: JWT token válido por 24 horas
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          type: object
          properties:
            username:
              type: string
              example: "admin"
            role:
              type: string
              example: "admin"
        expiresIn:
          type: string
          example: "24h"
    AdminStatsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            overview:
              type: object
              properties:
                totalAppointments:
                  type: integer
                  example: 31
                pendingAppointments:
                  type: integer
                  example: 25
                completedAppointments:
                  type: integer
                  example: 6
                successRate:
                  type: integer
                  description: Porcentaje de citas completadas
                  example: 19
            rds:
              type: object
              properties:
                total:
                  type: integer
                  example: 31
                pe:
                  type: integer
                  example: 25
                cl:
                  type: integer
                  example: 6
            performance:
              type: object
              properties:
                avgProcessingTime:
                  type: string
                  example: "~250ms"
                systemHealth:
                  type: string
                  example: "Healthy - RDS Only"
                lastUpdate:
                  type: string
                  format: date-time
    AdminAppointmentsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            rds:
              type: object
              properties:
                pe:
                  type: object
                  properties:
                    appointments:
                      type: array
                      items:
                        $ref: '#/components/schemas/DetailedAppointment'
                    count:
                      type: integer
                      example: 25
                cl:
                  type: object
                  properties:
                    appointments:
                      type: array
                      items:
                        $ref: '#/components/schemas/DetailedAppointment'
                    count:
                      type: integer
                      example: 6
    DetailedAppointment:
      type: object
      properties:
        appointment_id:
          type: string
          format: uuid
        insured_id:
          type: string
        country_iso:
          type: string
          enum: [PE, CL]
        medic_name:
          type: string
          example: "Dr. Juan Pérez"
        specialty_name:
          type: string
          example: "Cardiología"
        center_name:
          type: string
          example: "Clínica Internacional"
        center_address:
          type: string
          example: "Av. 28 de Julio 1234"
        center_city:
          type: string
          example: "Lima"
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        appointment_date:
          type: string
          format: date-time
    AdminMedicsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            pe:
              type: array
              items:
                $ref: '#/components/schemas/MedicInfo'
            cl:
              type: array
              items:
                $ref: '#/components/schemas/MedicInfo'
            summary:
              type: object
              properties:
                totalMedics:
                  type: integer
                  example: 6
                byCountry:
                  type: object
                  properties:
                    PE:
                      type: integer
                      example: 3
                    CL:
                      type: integer
                      example: 3
    MedicInfo:
      type: object
      properties:
        medic_id:
          type: integer
        name:
          type: string
          example: "Dr. Juan Pérez"
        specialty_id:
          type: integer
        center_id:
          type: integer
        specialty_name:
          type: string
          example: "Cardiología"
        center_name:
          type: string
          example: "Clínica Internacional"
        center_address:
          type: string
        center_city:
          type: string
    AdminCentersResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            pe:
              type: array
              items:
                $ref: '#/components/schemas/Center'
            cl:
              type: array
              items:
                $ref: '#/components/schemas/Center'
    AdminSpecialtiesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            pe:
              type: array
              items:
                $ref: '#/components/schemas/Specialty'
            cl:
              type: array
              items:
                $ref: '#/components/schemas/Specialty'
    Center:
      type: object
      properties:
        center_id:
          type: integer
        name:
          type: string
          example: "Clínica Internacional"
        address:
          type: string
          example: "Av. 28 de Julio 1234"
        city:
          type: string
          example: "Lima"
        country_iso:
          type: string
          enum: [PE, CL]
    Specialty:
      type: object
      properties:
        specialty_id:
          type: integer
        name:
          type: string
          example: "Cardiología"

    # ========== ERROR SCHEMAS ==========
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Descripción del error
          example: "Invalid request parameters"
        message:
          type: string
          description: Mensaje detallado del error
          example: "insuredId must be exactly 5 digits"
        details:
          type: object
          description: Detalles adicionales del error (opcional)

# ============= EJEMPLO =============
...